+++ src/lib/components/input/index.tsx
@@ -3,7 +3,7 @@ import { Input as UIInput } from '../../../base/ui/input'
 import { cn } from '../../../util/cn'
 
 export interface InputProps extends ComponentProps<typeof UIInput> {
-  className?: string;
+  className?: string
 }
 
 export const Input = ({ className, ...props }: InputProps) => {