+++ src/lib/components/input/input.stories.tsx
@@ -27,14 +27,13 @@ export const Password = Template.bind({})
 Password.args = {
   type: 'password',
   placeholder: 'Password',
-
 }
 
 export const Email = Template.bind({})
 Email.args = {
   type: 'email',
   placeholder: 'example@address.com',
-  className: "invalid:ring-red-500 invalid:border-red-500 invalid:focus:ring-red-500 invalid:focus:border-red-500",
+  className: 'invalid:ring-red-500 invalid:border-red-500 invalid:focus:ring-red-500 invalid:focus:border-red-500',
 }
 
 export const Number = Template.bind({})